updating stepper with swipable view
